show proposals in admin and allow to create talk from them

+++ app/com/humantalks/internal/persons/PersonDbService.scala
@@ -1,13 +1,15 @@
 package com.humantalks.internal.persons
 
+import com.humantalks.exposed.proposals.{ Proposal, ProposalRepository }
 import com.humantalks.internal.talks.{ Talk, TalkRepository }
 import global.infrastructure.DbService
 import play.api.libs.json.{ Json, JsObject }
 import reactivemongo.api.commands.WriteResult
 
-import scala.concurrent.{ ExecutionContext, Future }
+import scala.concurrent.Future
+import scala.concurrent.ExecutionContext.Implicits.global
 
-case class PersonDbService(personRepository: PersonRepository, talkRepository: TalkRepository) extends DbService[Person, Person.Id, Person.Data, Person.Id] {
+case class PersonDbService(personRepository: PersonRepository, talkRepository: TalkRepository, proposalRepository: ProposalRepository) extends DbService[Person, Person.Id, Person.Data, Person.Id] {
   val name = personRepository.name
 
   def find(filter: JsObject = Json.obj(), sort: JsObject = personRepository.defaultSort): Future[List[Person]] = personRepository.find(filter, sort)
@@ -18,13 +20,17 @@ case class PersonDbService(personRepository: PersonRepository, talkRepository: T
   def update(elt: Person, data: Person.Data, by: Person.Id): Future[WriteResult] = personRepository.update(elt, data, by)
   def setRole(id: Person.Id, role: Option[Person.Role.Value]): Future[WriteResult] = personRepository.setRole(id, role)
 
-  def delete(id: Person.Id)(implicit ec: ExecutionContext): Future[Either[List[Talk], WriteResult]] = {
-    talkRepository.findForPerson(id).flatMap { talks =>
-      if (talks.isEmpty) {
-        personRepository.delete(id).map(r => Right(r))
-      } else {
-        Future(Left(talks))
-      }
+  def delete(id: Person.Id): Future[Either[(List[Talk], List[Proposal]), WriteResult]] = {
+    (for {
+      talks <- talkRepository.findForPerson(id)
+      proposals <- proposalRepository.findForPerson(id)
+    } yield (talks, proposals)).flatMap {
+      case (talks, proposals) =>
+        if (talks.isEmpty && proposals.isEmpty) {
+          personRepository.delete(id).map(r => Right(r))
+        } else {
+          Future.successful(Left((talks, proposals)))
+        }
     }
   }
 }
